Started User Input Parsing Subsystem Development (with Bug Fixes)
Started development on the User Input Parsing Subsystem.  Currently, the program should update an actual PWM module to match the frequency and duty cycle of the input signal it receives.  This is one step above the example because the example was only an approximation of a PWM signal, whereas this program actually uses the PWM modules within the PIC.

Added documentation on operations that are potentially not self-explanatory within Input_Capture_Example.c