Fixed Error in Initialization of Port A Pins
All Port A pins on this PIC (using our current configuration) are pulled to ground, except RA0 and RA1 (used by the motor driver module).

Due to this, the rest of the pins in Port A need to be configured as inputs so that the PIC does not inadvertantly drive a high voltage through the pin, which would short it out.