Added Initialization of Pins for Motor Driver
Pin RA0 will be used to control the direction of the motor.  According to the motor driver data sheet, an input high will rotate the motor in the forward direction, and a low in the backward direction.

Pin RA1 will be used to prevent the motor driver from entering sleep mode.  The sleep mode uses inverse logic, so a high is required to prevent the motor driver from entering sleep mode, and a low would cause the driver to enter sleep mode.

Both are set to high by default because initially the motor driver should be on and moving in the forward direction.