Simplified IC_Buffer Struct and Implemented Atomic Update Operation for Input Compare Modules
I found a way to update all three values at once atomically within the interrupt using the PIC's built in ICBUF register, which means that the only thing that needs to be tracked in the IC_Buffer struct is the current period's values.

In addition, the IC interrupt was modified to atomically store all 3 values in one phase, eliminating race conditions without the need for a complex IC_Buffer struct and associated functions.