Bug Fixes
Fixed an issue where the IC_Buffers were initialized in the header file, causing "multiple definition" compilation errors.  These initializations were moved to the top of the InputCapture.c file.

Now the PWM dependency sets the state of the ANS and TRIS registers for the pins it will be using in each module, preventing a situation where a pin that was set for input starts getting used for output (which would cause undefined behavior).  Each pin is now defined to be a digital output pin before PWM generation starts.

Fixed an issue where the definition of the Update function inside the IC_Module struct was "const IC_Module*" instead of "IC_Module*", causing compilation errors.