Huge Overhaul of the Input Capture Dependency and Quality of Life Changes for the PWM Dependencies
Moved all precompiler variables to the InputCapture.c and PWM.c source files rather than the headers to prevent pollution of the namespace when including the dependencies.

Added failsafes in the PWM and Input Capture module code to set the relevent pins to digital mode and input or output (for IC or PWM, respectively) just in case the user did not set these pins correctly prior to initializing the module.

Changed the clock source of all Input Capture modules from Fcy (the system clock) to Timer1 with a prescaler of 64.  This is because the frequency of the system clock was too high to resolve input signals that had a frequency less than ~250hz.  The prescaler of 64 (in reference to the system clock) allows lower frequency input signals to be interpreted without error (but decreases the maximum frequency to ~6kHz).

Fixed an error in the Input Capture modules where frequency of the input signal was being calculated incorrectly.  First, fullPeriodClockCycles was not converted to a double before dividing by the clock frequency (resulting in severe rounding errrors).  In addition, the clock frequency had to be changed from Fcy (its original base) to "Fcy/Timer1Prescaler" to account for the fact that the Timer1 counter register increments more slowly because of the prescaler of 1:64.

Finally, a new test file has been added that is intended to show that the wireles controller input can be captured by the PIC's Input Capture modules and analyzed to determine its duty cycle % and frequency.  This is the last major piece of basic functionality that had to be completed before the input signal could be interpreted by the PIC and used to generate output signals.