Altered PWM.c and Added Kill Switch Relay Control to the Final Project
Altered the PWM.c implementation so that the OC_Module_1 uses the prescaler on Timer1 to allow the PIC to use smaller frequencies (so that I can use a 50 Hz frequency to run the propulsion engine throttle control servo).

Added kill switch relay control to the final project driver.  This allows the PIC to control whether the engines can continue to run or not.

Fine tuned the propulsion engine throttle control servo so that 180 degrees of rotation can be achieved.  This may be decreased to ~30 degrees depending on the outcome of integration with the propulsion system.  However, for now, I have fine tuned it such that decreasing the range of motion is easier to accomplish.