Refactored Propulsion System Variables and Updated Turning Subsystem to Use PWM Instead of Static Pins
The 50% duty cycle used for the turning PWM signal is a placeholder.  This may need to change in future development.

Pin A2 = Enable Pin on the stepper motor
Pin A3 = Direction (CW/CCW) Pin on the stepper motor
PWM Module #2 (Pin P1) = signal to the stepper motor to begin turning