Changed PWM_Module #2 and the Function of the Propulsion Engine Directional Stepper Motor
Changed the PWM_Module 2 to use Timer1's prescaler of 64 to allow for smaller frequency PWM signals.

Updated the main driver such that the number of pulses sent to the stepper motor has a 1:1 relationship to the number of steps taken by the motor.  The base configuration is currently a PWM of 400Hz coupled with a 30 RPM max speed on the stepper motor, with each PWM pulse moving the stepper by +/-1 count.