Fixed Errors with Stepper Motor Control
Input Capture Module #4 is used to monitor the number of pulses sent to the stepper motor via the PWM Module #2.  This count is used to know when the user has moved the stepper too far in one direction and prevents further motion in that direction.

Fixed various issues (like changing Pin A2 unnecessarily and several incorrect if statements in the IC Module #4 Interrupt) that was causing the PIC to send inappropriate signals to the stepper motor, which caused it to malfunction.